Critical_Audit_Matter/requirements.txt
flask
pandas
requests
edgartools
google-generativeai
langchain
langchain-core
